Got a minimal example of searching working.
There are two issues: first, it is very slow, because the searching
does not filter out components. Secondly, there will eventually be a
need to allow the user to select categories based on the filtered
results, as if there was no filtering at all (i.e. tick a box, and
have it select all the items below it that match the search
terms). This is not compatible with the current way the selection
works, and will require refactoring the selection logic (which is in
need of fixing anyway). Looking at that is the next step